+++ src/main/java/br/com/alura/services/relatorios/Relatorio.java
@@ -1,0 +1,17 @@
+package br.com.alura.services.relatorios;
+
+import java.util.Set;
+
+import br.com.alura.services.Fechamento;
+
+public abstract class Relatorio {
+    Fechamento fechamento;
+    
+    public Relatorio(Fechamento fechamento) {
+        this.fechamento = fechamento;
+    }
+
+    public void imprimeTodosOsRelatorio(Set<Relatorio> relatorios) {
+        relatorios.iterator().forEachRemaining(Relatorio::toString);;
+    }
+}
