+++ src/main/java/br/com/alura/services/relatorios/RelatorioSintetico.java
@@ -1,0 +1,9 @@
+package br.com.alura.services.relatorios;
+
+public class RelatorioSintetico {
+
+
+  public RelatorioSintetico() {
+  
+  }
+}
